Ajuste do filtro por usuário logado

+++ src/Back-End/ProFin.Core/Services/TransactionService.cs
@@ -2,7 +2,7 @@
 using ProFin.Core.Interfaces.Repositories;
 using ProFin.Core.Interfaces.Services;
 using ProFin.Core.Models;
-using ProFin.Core.Models.Validations;
+using ProFin.Core.Models.Validations.Transaction;
 using System.Linq.Expressions;
 
 namespace ProFin.Core.Services
@@ -18,15 +18,29 @@ namespace ProFin.Core.Services
 
         public async Task Insert(FinancialTransaction transactionEntity)
         {
+            if (!_userService.IsAuthenticated())
+            {
+                Notifie("Categoria só pode ser adcionada por um usuário autenticado");
+                return;
+            }
+
             if (!ExecuteValidation(new TransactionEntityValidation(), transactionEntity)) return;
 
-
+            transactionEntity.SetUset(_userService.GetId().Value);
             await _transactionRepository.Add(transactionEntity);
         }
 
         public async Task Update(FinancialTransaction transactionEntity)
         {
-            if (!ExecuteValidation(new TransactionEntityValidation(), transactionEntity)) return;
+
+            if (!_userService.IsAuthenticated())
+            {
+                Notifie("Categoria só pode ser alterada por um usuário autenticado");
+                return;
+            }
+
+            if (!ExecuteValidation(new UpdateTransactionValidation(_userService.GetId().GetValueOrDefault()),
+                transactionEntity)) return;
 
 
             await _transactionRepository.Update(transactionEntity);
@@ -35,6 +49,14 @@ namespace ProFin.Core.Services
         public async Task Delete(Guid id)
         {
             var entity = await _transactionRepository.GetById(id);
+
+            if (!_userService.IsAuthenticated())
+            {
+                Notifie("Categoria só pode ser alterada por um usuário autenticado");
+                return;
+            }
+
+
             if (entity != null && entity.CreatedDate != DateTime.MinValue)
                 await _transactionRepository.Delete(entity);
             else
@@ -43,7 +65,17 @@ namespace ProFin.Core.Services
 
         public async Task<IEnumerable<FinancialTransaction>> GetSince(DateTime startedDate)
         {
-            Expression<Func<FinancialTransaction, bool>> filter = x => x.CreatedDate.Date >= startedDate;
+            if (_userService.IsAuthenticated() == false)
+                return Enumerable.Empty<FinancialTransaction>();
+
+            if (_userService.IsAdmin())
+            {
+                Expression<Func<FinancialTransaction, bool>> filters = x => x.CreatedDate.Date >= startedDate;
+                return await _transactionRepository.GetAll(includes: "CategoryFinancialTransaction", filters);
+            }
+
+
+            Expression<Func<FinancialTransaction, bool>> filter = x => x.CreatedDate.Date >= startedDate && x.UserId == _userService.GetId().Value;
             var data = await _transactionRepository.GetAll(includes: "CategoryFinancialTransaction", filter);
             return data;
         }
